complete training loop, dtype bug fixes

+++ core/pointnet.py
@@ -6,24 +6,24 @@ import torch.utils.data
 from torch.autograd import Variable
 import numpy as np
 import torch.nn.functional as F
-torch.set_default_dtype(torch.float64)
+torch.set_default_dtype(torch.float32)
 
 class STN3d(nn.Module):
     def __init__(self):
         super().__init__()
-        self.conv1 = torch.nn.Conv1d(3, 64, 1)
-        self.conv2 = torch.nn.Conv1d(64, 128, 1)
-        self.conv3 = torch.nn.Conv1d(128, 1024, 1)
-        self.fc1 = nn.Linear(1024, 512)
-        self.fc2 = nn.Linear(512, 256)
-        self.fc3 = nn.Linear(256, 9)
+        self.conv1 = torch.nn.Conv1d(3, 64, 1,dtype=torch.float32)
+        self.conv2 = torch.nn.Conv1d(64, 128, 1,dtype=torch.float32)
+        self.conv3 = torch.nn.Conv1d(128, 1024, 1,dtype=torch.float32)
+        self.fc1 = nn.Linear(1024, 512,dtype=torch.float32)
+        self.fc2 = nn.Linear(512, 256,dtype=torch.float32)
+        self.fc3 = nn.Linear(256, 9,dtype=torch.float32)
         self.relu = nn.ReLU()
 
-        self.bn1 = nn.BatchNorm1d(64)
-        self.bn2 = nn.BatchNorm1d(128)
-        self.bn3 = nn.BatchNorm1d(1024)
-        self.bn4 = nn.BatchNorm1d(512)
-        self.bn5 = nn.BatchNorm1d(256)
+        self.bn1 = nn.BatchNorm1d(64,dtype=torch.float32)
+        self.bn2 = nn.BatchNorm1d(128,dtype=torch.float32)
+        self.bn3 = nn.BatchNorm1d(1024,dtype=torch.float32)
+        self.bn4 = nn.BatchNorm1d(512,dtype=torch.float32)
+        self.bn5 = nn.BatchNorm1d(256,dtype=torch.float32)
 
 
     def forward(self, x):
@@ -38,7 +38,7 @@ class STN3d(nn.Module):
         x = F.relu(self.bn5(self.fc2(x)))
         x = self.fc3(x)
 
-        iden = Variable(torch.from_numpy(np.array([1,0,0,0,1,0,0,0,1]).astype(np.float64))).view(1,9).repeat(batchsize,1)
+        iden = Variable(torch.from_numpy(np.array([1,0,0,0,1,0,0,0,1]).astype(np.float32))).view(1,9).repeat(batchsize,1)
         if x.is_cuda:
             iden = iden.cuda()
         x = x + iden
@@ -49,19 +49,19 @@ class STN3d(nn.Module):
 class STNkd(nn.Module):
     def __init__(self, k=64):
         super().__init__()
-        self.conv1 = torch.nn.Conv1d(k, 64, 1)
-        self.conv2 = torch.nn.Conv1d(64, 128, 1)
-        self.conv3 = torch.nn.Conv1d(128, 1024, 1)
-        self.fc1 = nn.Linear(1024, 512)
-        self.fc2 = nn.Linear(512, 256)
-        self.fc3 = nn.Linear(256, k*k)
+        self.conv1 = torch.nn.Conv1d(k, 64, 1, dtype=torch.float32)
+        self.conv2 = torch.nn.Conv1d(64, 128, 1, dtype=torch.float32)
+        self.conv3 = torch.nn.Conv1d(128, 1024, 1, dtype=torch.float32)
+        self.fc1 = nn.Linear(1024, 512, dtype=torch.float32)
+        self.fc2 = nn.Linear(512, 256, dtype=torch.float32)
+        self.fc3 = nn.Linear(256, k*k, dtype=torch.float32)
         self.relu = nn.ReLU()
 
-        self.bn1 = nn.BatchNorm1d(64)
-        self.bn2 = nn.BatchNorm1d(128)
-        self.bn3 = nn.BatchNorm1d(1024)
-        self.bn4 = nn.BatchNorm1d(512)
-        self.bn5 = nn.BatchNorm1d(256)
+        self.bn1 = nn.BatchNorm1d(64, dtype=torch.float32)
+        self.bn2 = nn.BatchNorm1d(128,dtype=torch.float32)
+        self.bn3 = nn.BatchNorm1d(1024,dtype=torch.float32)
+        self.bn4 = nn.BatchNorm1d(512,dtype=torch.float32)
+        self.bn5 = nn.BatchNorm1d(256,dtype=torch.float32)
 
         self.k = k
 
@@ -77,7 +77,7 @@ class STNkd(nn.Module):
         x = F.relu(self.bn5(self.fc2(x)))
         x = self.fc3(x)
 
-        iden = Variable(torch.from_numpy(np.eye(self.k).flatten().astype(np.float64))).view(1,self.k*self.k).repeat(batchsize,1)
+        iden = Variable(torch.from_numpy(np.eye(self.k).flatten().astype(np.float32))).view(1,self.k*self.k).repeat(batchsize,1)
         if x.is_cuda:
             iden = iden.cuda()
         x = x + iden
@@ -88,12 +88,12 @@ class PointNetfeat(nn.Module):
     def __init__(self, global_feat = True, feature_transform = False):
         super().__init__()
         self.stn = STN3d()
-        self.conv1 = torch.nn.Conv1d(3, 64, 1)
-        self.conv2 = torch.nn.Conv1d(64, 128, 1)
-        self.conv3 = torch.nn.Conv1d(128, 1024, 1)
-        self.bn1 = nn.BatchNorm1d(64)
-        self.bn2 = nn.BatchNorm1d(128)
-        self.bn3 = nn.BatchNorm1d(1024)
+        self.conv1 = torch.nn.Conv1d(3, 64, 1, dtype=torch.float32)
+        self.conv2 = torch.nn.Conv1d(64, 128, 1, dtype=torch.float32)
+        self.conv3 = torch.nn.Conv1d(128, 1024, 1, dtype=torch.float32)
+        self.bn1 = nn.BatchNorm1d(64, dtype=torch.float32)
+        self.bn2 = nn.BatchNorm1d(128, dtype=torch.float32)
+        self.bn3 = nn.BatchNorm1d(1024, dtype=torch.float32)
         self.global_feat = global_feat
         self.feature_transform = feature_transform
         if self.feature_transform:
@@ -183,12 +183,12 @@ class PointNetFeatureExtractor(nn.Module):
         self.feature_transform=feature_transform
         # global_feat = false concatenates nx1024 with nx64 from 2nd T-net (desired behaviour)
         self.feat = PointNetfeat(global_feat=False, feature_transform=feature_transform)
-        self.conv1 = torch.nn.Conv1d(1088, 512, 1)
-        self.conv2 = torch.nn.Conv1d(512, 256, 1)
-        self.conv3 = torch.nn.Conv1d(256, 128, 1)
-        self.bn1 = nn.BatchNorm1d(512)
-        self.bn2 = nn.BatchNorm1d(256)
-        self.bn3 = nn.BatchNorm1d(128)
+        self.conv1 = torch.nn.Conv1d(1088, 512, 1, dtype=torch.float32)
+        self.conv2 = torch.nn.Conv1d(512, 256, 1, dtype=torch.float32)
+        self.conv3 = torch.nn.Conv1d(256, 128, 1, dtype=torch.float32)
+        self.bn1 = nn.BatchNorm1d(512, dtype=torch.float32)
+        self.bn2 = nn.BatchNorm1d(256, dtype=torch.float32)
+        self.bn3 = nn.BatchNorm1d(128, dtype=torch.float32)
 
     def forward(self, x):
         batchsize = x.size()[0]
@@ -206,7 +206,8 @@ def feature_transform_regularizer(trans):
     I = torch.eye(d)[None, :, :]
     if trans.is_cuda:
         I = I.cuda()
-    loss = torch.mean(torch.norm(torch.bmm(trans, trans.transpose(2,1)) - I, dim=(1,2)))
+    
+    loss = torch.mean(torch.norm(torch.bmm(trans, trans.transpose(2,1)) - I, dim=(1,2), dtype=torch.float32), dtype=torch.float32)
     return loss
 
 if __name__ == '__main__':
